Extract models into app/models.py
- Created app/models.py with User and Scenario models
- Moved db initialization to app/__init__.py for proper package structure
- Added datetime import to models.py for Scenario timestamps
- Updated app.py to import User and Scenario from app.models
- Updated app.py to import db from app package
- Removed old model class definitions from app.py
- All features tested and working (login, calculate, save, history, delete)

Relates to #21
